Clean paths and integrate battle screen
