--- src/# Code Citations.md
@@ -1,16 +1,0 @@
-# Code Citations
-
-## License: MIT
-https://github.com/Alsamraeammar/GAME/tree/4065f7a623cb6a6eb0a5272d4108c6f5c93aa5a4/game.py
-
-```
-self.clock.tick(60)
-        pygame.quit()
-
-    def handle_events(self):
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                self.running = False
-            elif event.type == pygame.KEYDOWN:
-```
-
