+++ src/screens/end.py
@@ -2,10 +2,12 @@ import pygame
 from game import ScreenNames
 
 class EndScreen:
-    def __init__(self, app): self.app = app
+    def __init__(self, screen, game):
+        self.screen = screen
+        self.game = game
     def on_enter(self): pass
     def handle_event(self, event): pass
     def update(self, dt): pass
     def draw(self):
-        self.app.screen.fill((100,0,0))
+        self.screen.fill((100,0,0))
         pass
